requirements.txt
python-dotenv
dspy
numpy
pillow
pydantic
ffmpeg